# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoyaStar Backend API",
        "description": "LMS platform for managing students, teachers, schools, and curriculum.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "List all videos",
                "operationId": "555e20ea8f656d0a1575e693caf51c39",
                "parameters": [
                    {
                        "name": "lesson_id",
                        "in": "query",
                        "description": "Filter by lesson ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (pending_upload, ready)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort column (e.g. name, created_at)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 20)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of videos"
                    }
                },
                "security": [
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Create video",
                "operationId": "dffe90eee1548387a8703bba957be3da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video created with upload link"
                    }
                }
            }
        },
        "/videos/{video}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get video by ID",
                "operationId": "a1af100e9eb8ec89fdd5c7199ac8e09a",
                "parameters": [
                    {
                        "name": "video",
                        "in": "path",
                        "description": "Video ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video data"
                    }
                },
                "security": [
                    {}
                ]
            },
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update video",
                "operationId": "ec42a8ab546f092484a1c73e564ee0de",
                "responses": {
                    "200": {
                        "description": "Video updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Videos"
                ],
                "summary": "Delete video",
                "operationId": "22d78d0da83052bf7714131712512d03",
                "responses": {
                    "200": {
                        "description": "Video deleted"
                    }
                }
            }
        },
...
        "/admin/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "List questions",
                "operationId": "c052ebc3a7aedb9c51b4a5d20e165fc9",
                "parameters": [
                    {
                        "name": "topic_id",
                        "in": "query",
                        "description": "Filter by single topic ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "topic_ids",
                        "in": "query",
                        "description": "Filter by multiple topic IDs (comma-separated)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by type (regular, open)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "difficulty_level",
                        "in": "query",
                        "description": "Filter by difficulty level",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by question text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort column (e.g. created_at)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default 20)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions"
                    }
                }
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create question",
                "operationId": "46b31cc056835f2f7611717d5096188b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "topic_id",
                                    "difficulty_level"
                                ],
                                "properties": {
                                    "question": {
                                        "description": "Question content (type + content)",
                                        "type": "array"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "regular",
                                            "open"
                                        ]
                                    },
                                    "explanation": {
                                        "type": "array",
                                        "nullable": true
                                    },
                                    "difficulty_level": {
                                        "type": "string"
                                    },
                                    "topic_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created"
                    }
                }
            }
        },
        "/admin/questions/{question}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get question by ID",
                "operationId": "ba0bfede1775d79f79bf4de9fbc51023",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question data"
                    }
                }
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update question",
                "operationId": "b3932538310fd3ff53da5ff7a647796d",
                "responses": {
                    "200": {
                        "description": "Question updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete question",
                "operationId": "6d8e801bd1605fe0901c2df020609788",
                "responses": {
                    "200": {
                        "description": "Question deleted"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "bearer",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}